Changed some properties in the models. Added initial DataHolder.java with sample info...

+++ src/main/java/football/clubs/repository/StadiumRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import java.util.Optional;
 
-public interface StadiumRepository extends JpaRepository<Stadium,Long> {
+public interface StadiumRepository extends JpaRepository<Stadium,Integer> {
 
 }
