+++ src/main/java/football/clubs/service/ClubService.java
@@ -9,7 +9,7 @@ public interface ClubService {
 
     public Club save(Club club);
     public List<Club> findAll();
-    public Optional<Club> findById(Long id);
+    public Optional<Club> findById(Integer id);
     public Optional<Club> findByClubName(String Name);
     public Club getClubDetailsByLink(String clubLink);
     public Club getClubDetailsByName(String Name);
