+++ src/main/java/football/clubs/service/CountryService.java
@@ -9,7 +9,7 @@ import java.util.Optional;
 public interface CountryService {
     public Country save(Country country);
     public List<Country> findAll();
-    public Optional<Country> findById(Long id);
+    public Optional<Country> findById(Integer id);
     public Optional<Country> findByCountryName(String Name);
     public Country getCountryDetailsByLink(String countryLink);
     public Country getCountryDetailsByName(String Name);
