+++ src/main/java/football/clubs/service/StadiumService.java
@@ -9,6 +9,6 @@ import java.util.Optional;
 public interface StadiumService {
     public Stadium save(Stadium stadium);
     public List<Stadium> findAll();
-    public Optional<Stadium> findById(Long id);
+    public Optional<Stadium> findById(Integer id);
     public Stadium getStadiumDetailsByLink(String stadiumLink);
 }
